Extract the MovieCard component from App component.
